+++ src/main/java/de/uib/configed/gui/DepotListCellRenderer.java
@@ -7,22 +7,37 @@
 package de.uib.configed.gui;
 
 import java.awt.Component;
+import java.awt.image.BufferedImage;
 import java.util.Map;
 
 import javax.swing.DefaultListCellRenderer;
+import javax.swing.ImageIcon;
 import javax.swing.JList;
 import javax.swing.UIManager;
 
 import de.uib.configed.Configed;
+import de.uib.configed.ConfigedMain;
+import de.uib.opsicommand.ServerFacade;
 import de.uib.opsidatamodel.serverdata.OpsiServiceNOMPersistenceController;
 import de.uib.opsidatamodel.serverdata.PersistenceControllerFactory;
 import de.uib.utilities.logging.Logging;
+import utils.Utils;
 
 public class DepotListCellRenderer extends DefaultListCellRenderer {
 	Map<String, Map<String, Object>> extendedInfo;
 
+	private ConfigedMain configedMain;
+
 	private OpsiServiceNOMPersistenceController persistenceController = PersistenceControllerFactory
 			.getPersistenceController();
+
+	private ImageIcon configServerConnectedIcon = Utils.createImageIcon("bootstrap/check_circle_blue.png", "");
+	private ImageIcon configServerDisconnectedIcon = Utils.createImageIcon("bootstrap/circle_blue.png", "");
+	private ImageIcon connectedIcon = Utils.createImageIcon("bootstrap/check_green.png", "");
+
+	public DepotListCellRenderer(ConfigedMain configedMain) {
+		this.configedMain = configedMain;
+	}
 
 	public void setInfo(Map<String, Map<String, Object>> extendedInfo) {
 		Logging.debug(this, "setInfo " + extendedInfo);
@@ -47,6 +62,10 @@ public class DepotListCellRenderer extends DefaultListCellRenderer {
 			tooltipText = "" + extendedInfo.get(value).get("description");
 		}
 
+		if (ServerFacade.isOpsi43()) {
+			setConnectionIcon(value);
+		}
+
 		String depot = (String) value;
 		if (!persistenceController.getUserRolesConfigDataService().hasDepotPermission(depot)) {
 			setEnabled(false);
@@ -58,5 +77,20 @@ public class DepotListCellRenderer extends DefaultListCellRenderer {
 		}
 
 		return this;
+	}
+
+	private void setConnectionIcon(Object value) {
+		if (configedMain.getConnectedClientsByMessagebus().contains(value)) {
+			setIcon(connectedIcon);
+		} else if (value != null && value.equals(persistenceController.getHostInfoCollections().getConfigServer())) {
+			if (configedMain.getMessagebus().isConnected()) {
+				setIcon(configServerConnectedIcon);
+			} else {
+				setIcon(configServerDisconnectedIcon);
+			}
+		} else {
+			BufferedImage emptyImage = new BufferedImage(connectedIcon.getIconWidth(), 1, BufferedImage.TYPE_INT_ARGB);
+			setIcon(new ImageIcon(emptyImage));
+		}
 	}
 }